requirements.txt
selenium==4.15.0
pandas==2.1.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
openpyxl==3.1.2
lxml==4.9.3
webdriver-manager==4.0.1
fake-useragent==1.4.0